# docs/v3/api_spec/crs_converter_openapi.json
{
  "swagger": "2.0",
  "info": {
    "description": "Provides conversion/transformation services from a source to a target CRS. A coordinate operation is a conversion if the source and target CRS share the same datum; the operation is called transformation if datum transformations are involved, i.e. source and target CRS do not share the same datum.",
    "version": "1.0.0",
    "title": "CRS Conversion Service"
  },
  "host": "az-osdu1.evd.csp.slb.com",
...
  "basePath": "/api/crs/v1",
  "security": [
    {
      "Bearer": []
    },
    {
      "google_id_token": []
    }
  ],
...
  "schemes": [
    "https"
  ],
  "paths": {
    "/v3/convertTrajectory": {
      "post": {
        "tags": [
          "Trajectory Computation and Conversion"
        ],
        "summary": "Convert trajectory stations",
        "description": "Convert a list of trajectory stations, given the unit and spatial context and a reference point in 3D where MD==0.",
        "operationId": "convertTrajectory",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/ConvertTrajectoryRequest"
            }
          },
          {
            "in": "header",
            "name": "data-partition-id",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Trajectory conversion performed successfully",
            "schema": {
              "$ref": "#/definitions/ConvertTrajectoryResponse"
            }
          },
          "400": {
            "description": "Bad input format",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "503": {
            "description": "CRS-converter overloaded; try again later. See HttpHeader 'Retry-After'.",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          },
          {
            "google_id_token": []
          }
        ]
      }
    },
    "/v3/convert": {
      "post": {
        "tags": [
          "CRS Point Conversion"
        ],
        "summary": "Convert a list of points",
        "description": "Convert a list of points",
        "operationId": "convertPoint",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/ConvertPointsRequest"
            }
          },
          {
            "in": "header",
            "name": "data-partition-id",
            "required": false
          }
        ],

        "responses": {
          "200": {
            "description": "Conversion performed successfully",
            "schema": {
              "$ref": "#/definitions/ConvertPointsResponse"
            }
          },
          "400": {
            "description": "Bad input format",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "503": {
            "description": "CRS-converter overloaded; try again later. See HttpHeader 'Retry-After'.",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          },
          {
            "google_id_token": []
          }
        ]
      }
    },
    "/v3/convertGeoJson": {
      "post": {
        "tags": [
          "CRS Point Conversion"
        ],
        "summary": "Convert a GeoJSON or AnyCrsGeoJson structure",
        "description": "Convert a GeoJSON or AnyCrsGeoJson structure to a specified target CRS. GeoJSON is declared - by definition - to be in the context of WGS 84.",
        "operationId": "convertGeoJson",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/ConvertGeoJsonRequest"
            }
          },
          {
            "in": "header",
            "name": "data-partition-id",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Conversion performed successfully",
            "schema": {
              "$ref": "#/definitions/ConvertGeoJsonResponse"
            }
          },
          "400": {
            "description": "Bad input format",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "503": {
            "description": "CRS-converter overloaded; try again later. See HttpHeader 'Retry-After'.",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          },
          {
            "google_id_token": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
    "ConvertTrajectoryResponse": {
      "type": "object",
      "required": [
        "localCRS",
        "method",
        "stations",
        "trajectoryCRS",
        "unitDls",
        "unitZ"
      ],
      "properties": {
        "trajectoryCRS": {
          "type": "string",
          "example": "osdu:reference-data--CoordinateReferenceSystem:BoundCRS::osdu::toCRStrajectory23031024",
          "description": "Coordinate reference system for the reference point; typically the CRS is a projected CRS; if a geographic CRS is provided, the unitXY must be defined and the azimuthReference must be TrueNorth."
        },
        "unitXY": {
          "type": "string",
          "example": "osdu:reference-data--UnitOfMeasure:ms",
          "description": "The horizontal unit of the dx, dy in the output trajectory stations."
        },
        "unitZ": {
          "type": "string",
          "example": "osdu:reference-data--UnitOfMeasure:ms",
          "description": "The vertical unit of the dz in the output trajectory stations."
        },
        "unitDls": {
          "type": "string",
          "example": "{\"scaleOffset\":{\"scale\":5.72614583987641E-4,\"offset\":0.0},\"symbol\":\"deg/100ft\",\"baseMeasurement\":{\"ancestry\":\"Rotation_Per_Length\",\"type\":\"UM\"},\"type\":\"USO\"}",
          "description": "The unit of the dog leg severity (DLS) in the output trajectory stations."
        },
        "stations": {
          "type": "array",
          "description": "Computed trajectory stations.",
          "items": {
            "$ref": "#/definitions/TrajectoryStationOut"
          }
        },
        "localCRS": {
          "type": "string",
          "example": "osdu:reference-data--CoordinateReferenceSystem:BoundCRS::osdu::toCRStrajectory23031024",
            "description": "Coordinate Reference System for the local, True North oriented, true distance, engineering CRS with origin at the well's surface location."
        },
        "method": {
          "type": "string",
          "example": "AzimuthalEquidistant",
          "description": "The computation method used - 'AzimuthalEquidistant' (default) or 'LMP' (Lee's modified proposal SPE96813)."
        },
        "operationsApplied": {
          "type": "array",
          "description": "The list of operations, which have been applied to the points",
          "items": {
            "type": "string"
          }
        }
      },
      "description": "Trajectory response data structure; contains the context (crs, units)."
    },
    "Point": {
      "type": "object",
      "required": [
        "x",
        "y",
        "z"
      ],
      "properties": {
        "x": {
          "type": "number",
          "format": "double",
          "example": -61.04340628871454,
          "description": "x coordinate or longitude"
        },
        "y": {
          "type": "number",
          "format": "double",
          "example": 10.673103179456877,
          "description": "y coordinate or latitude"
        },
        "z": {
          "type": "number",
          "format": "double",
          "example": 0,
          "description": "z coordinate"
        }
      },
      "description": "Point representation for CRS operations"
    },
...
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "description": "Error message"
        }
      },
      "description": "Response resulting of a failed CRS convertion operation"
    },
    "ConvertTrajectoryRequest": {
      "type": "object",
      "required": [
        "azimuthReference",
        "inputStations",
        "method",
        "referencePoint",
        "trajectoryCRS",
        "unitZ"
      ],
      "properties": {
        "trajectoryCRS": {
          "type": "string",
          "example": "osdu:reference-data--CoordinateReferenceSystem:BoundCRS::osdu::toCRStrajectory23031024",
          "description": "Coordinate reference system for the reference point; typically the CRS is a projected CRS; if a geographic CRS is provided, the unitXY must be defined and the azimuthReference must be TrueNorth. The property value could be a persistable reference string or a record ID"
        },
        "azimuthReference": {
          "type": "string",
          "example": "TrueNorth",
          "description": "Azimuth reference for the input trajectory station azimuth values (TrueNorth or GridNorth)"
        },
        "unitXY": {
          "type": "string",
          "example": "osdu:reference-data--UnitOfMeasure:ms",
          "description": "The horizontal unit of the dx, dy in the input trajectory stations; the unit must be a length unit in persistable reference or a record ID, see example."
        },
        "unitZ": {
          "type": "string",
          "example": "osdu:reference-data--UnitOfMeasure:ms",
          "description": "The vertical unit of the dz in the input trajectory stations; the unit could be a persistable reference or a record ID"
        },
        "referencePoint": {
          "description": "The 3D reference point in the 'trajectoryCRS' where MD==0.",
          "$ref": "#/definitions/Point"
        },
        "inputStations": {
          "type": "array",
          "description": "The array of input trajectory stations",
          "items": {
            "$ref": "#/definitions/TrajectoryStationIn"
          }
        },
        "method": {
          "type": "string",
          "example": "AzimuthalEquidistant",
          "description": "The computation method - 'AzimuthalEquidistant' (default) or 'LMP' (Lee's modified proposal SPE96813)"
        },
        "inputKind": {
          "type": "string",
          "example": "MD_Incl_Azim",
          "description": "The kind of input; one of MD_Incl_Azim (default), MD_X_Y_Z, MD_dX_dY_dZ, X_Y_Z, dX_dY_dZ. MD stands for measured depth; MD_X_Y_Z/X_Y_Z stand for absolute coordinates in the reference CRS, MD_dX_dY_dZ/dX_dY_dZ stand for deviations relative to the reference point."
        },
        "interpolate": {
          "type": "boolean",
          "description": "Perform trajectory interpolation on demand; default is true."
        }
      },
      "description": "Input trajectory data structure; contains the context (crs, units, azimuth reference, method)",
      "example": {
        "trajectoryCRS": "osdu:reference-data--CoordinateReferenceSystem:BoundCRS::osdu::toCRStrajectory23031024",
        "azimuthReference": "TN",
        "unitXY": "osdu:reference-data--UnitOfMeasure:ms",
        "unitZ": "osdu:reference-data--UnitOfMeasure:ms",
        "referencePoint": {
          "x": 400000,
          "y": 6500000,
          "z": 0
        },
        "inputKind": "MD_Incl_Azim",
        "interpolate": true,
        "inputStations": [
          {
            "md": 0,
            "inclination": 0,
            "azimuth": 0
          },
          {
            "md": 1000,
            "inclination": 0,
            "azimuth": 0
          },
          {
            "md": 2000,
            "inclination": 90,
            "azimuth": 0
          },
          {
            "md": 3000,
            "inclination": 90,
            "azimuth": 0
          },
          {
            "md": 5000,
            "inclination": 90,
            "azimuth": 90
          },
          {
            "md": 6000,
            "inclination": 90,
            "azimuth": 90
          }
        ],
        "method": "AzimuthalEquidistant"
      }
    },
...
    "ConvertPointsResponse": {
      "type": "object",
      "properties": {
        "successCount": {
          "type": "integer",
          "format": "int32",
          "description": "Number of points successfully converted. If the number is less than the request array length conversion/transformation failures occurred."
        },
        "points": {
          "type": "array",
          "description": "Converted points; length and order of the array is the same as in the request. Points, which failed to convert, are returned as NaN.",
          "items": {
            "$ref": "#/definitions/Point"
          }
        },
        "operationsApplied": {
          "type": "array",
          "description": "The list of operations, which have been applied to the points",
          "items": {
            "type": "string"
          }
        }
      },
      "description": "Response of a CRS conversion/transformation operation"
    },
    "ConvertPointsRequest": {
      "type": "object",
      "required": [
        "fromCRS",
        "points",
        "toCRS"
      ],
      "properties": {
        "fromCRS": {
          "type": "string",
          "example": "osdu:reference-data--CoordinateReferenceSystem:BoundCRS::osdu::fromcrsConvert23031024",
          "description": "Source CRS as persistable reference string or a record ID"
        },
        "toCRS": {
          "type": "string",
          "example": "osdu:reference-data--CoordinateReferenceSystem:BoundCRS::osdu::toCRSconvert23031024",
          "description": "Target CRS as persistable reference string or a record ID"
        },
        "points": {
          "type": "array",
          "description": "List of points to be converted",
          "items": {
            "$ref": "#/definitions/Point"
          }
        }
      },
      "description": "Request to convert a set of points from a source CRS to a target CRS.",
      "example": {
        "fromCRS": "osdu:reference-data--CoordinateReferenceSystem:BoundCRS::osdu::fromcrsConvert23031024",
        "toCRS":"osdu:reference-data--CoordinateReferenceSystem:BoundCRS::osdu::toCRSconvert23031024",
        "points": [
          {
            "x": -61.04340628871454,
            "y": 10.673103179456877,
            "z": 0
          }
        ]
      }
    },
    "ConvertGeoJsonRequest": {
      "title": "Convert GeoJson/AnyCrsGeoJson Request",
      "description": "Request to convert a GeoJSON FeatureCollection or AnyCrsFeatureCollection from WGS 84 or AnyCrsFeatureCollection.persistableReferenceCrs to a target CRS.",
      "type": "object",
      "required": [
        "featureCollection",
        "toCRS"
      ],
      "properties": {
        "featureCollection": {
          "$ref": "#/definitions/AnyCrsGeoJsonFeatureCollection"
        },
        "toCRS": {
          "title": "GeoJson Request",
          "type": "string",
          "description": "The GeoJSON FeatureCollection or AnyCrsFeatureCollection structure to be converted/transformed. GeoJSON is always based on WGS 84; AnyCrsFeatureCollection carries the CRS context in the persistableReferenceCrs property. This value could be persistable reference string or a record ID"
        },
        "toUnitZ": {
          "type": "string",
          "description": "The vertical axis unit for this AnyCrsGeoJsonFeatureCollection as persistable reference string or a record ID)",
          "title": "Vertical Unit"
        }
      },
      "example": {
        "toCRS": "osdu:reference-data--CoordinateReferenceSystem:BoundCRS::osdu::toCRSGeojson23031024",
        "toUnitZ": "osdu:reference-data--UnitOfMeasure:ms",
        "featureCollection": {
          "features": [
            {
              "geometry": {
                "coordinates": [
                  313405.9477893702,
                  6544797.620047403,
                  6.561679790026246
                ],
                "bbox": null,
                "type": "AnyCrsPoint"
              },
              "bbox": null,
              "properties": {},
              "type": "AnyCrsFeature"
            }
          ],
          "bbox": null,
          "properties": {},
          "CoordinateReferenceSystemID": "osdu:reference-data--CoordinateReferenceSystem:BoundCRS::osdu::persistGeojson23031024",
          "VerticalUnitZ": "osdu:reference-data--UnitOfMeasure:ms",
          "type": "AnyCrsFeatureCollection"
        }
      }
    },
...
  }
}